RegisterTracer cannot be used in Go 1.4
