+++ tracer.go
@@ -1,20 +1,12 @@
 package proxy
 
 import (
-	"database/sql"
 	"database/sql/driver"
 	"fmt"
-	"log"
 	"runtime"
 	"strings"
 	"time"
 )
-
-type logger struct{}
-
-func (_ logger) Output(calldepth int, s string) error {
-	return log.Output(calldepth, s)
-}
 
 // Outputter is what is used by the tracing proxy created via `NewTraceProxy`.
 // Anything that implements a `log.Logger` style `Output` method will satisfy
@@ -154,19 +146,5 @@ func NewTraceProxyWithFilter(d driver.Driver, o Outputter, f Filter) *Proxy {
 				return nil
 			},
 		},
-	}
-}
-
-// RegisterTracer creates proxies that logs queries from the sql drivers already registered,
-// and registers the proxies as sql driver.
-// The proxies' names have suffix ":trace".
-func RegisterTracer() {
-	for _, driver := range sql.Drivers() {
-		if strings.HasSuffix(driver, ":trace") {
-			continue
-		}
-		db, _ := sql.Open(driver, "")
-		defer db.Close()
-		sql.Register(driver+":trace", NewTraceProxy(db.Driver(), logger{}))
 	}
 }
